img-txt_viewer/requirements.txt
numpy==2.1.3
Pillow==11.0.0
onnxruntime==1.20.0
onnxruntime-directml==1.20.1
git+https://github.com/Nenotriple/TkToolTip.git@main
git+https://github.com/Nenotriple/tkVideoPlayer.git@master
